% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapidxmlr-package.R
\docType{package}
\name{rapidxmlr-package}
\alias{rapidxmlr}
\alias{rapidxmlr-package}
\title{Provides XML parsing capability through the 'rapidxml' 'C++' header-only library}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\examples{
\dontrun{

# Include a dependency to rapidxmlr inside a cpp file
// [[Rcpp::export(rapidxmlr)]]
include <Rcpp.h>

}
}
