% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapidxmlr-package.R
\docType{package}
\name{rapidxmlr-package}
\alias{rapidxmlr}
\alias{rapidxmlr-package}
\title{Provides XML parsing capability through the 'rapidxml' 'C++' header-only library}
\description{
Provides XML parsing capability through the 'Rapidxml' 'C++' header-only library.
}
\examples{
\dontrun{

# Include a dependency to rapidxmlr inside a cpp file
// [[Rcpp::depends(rapidxmlr)]]
include <Rcpp.h>

}
}
\author{
\strong{Maintainer}: David Cooley \email{david.cooley.au@gmail.com}

Other contributors:
\itemize{
  \item Marcin Kalicinski (Author of c++ rapidxml library) [contributor]
}

}
